mgr/dashboard: Add dashboard to mon_command reference docs
(cherry picked from commit 6b029c09951d002fa6bf04390920f1232390fb2f)
