+++ src/pybind/mgr/dashboard/__init__.py
@@ -23,8 +23,6 @@ if 'UNITTEST' not in os.environ:
 
     mgr = _ModuleProxy()
 
-    # DO NOT REMOVE: required for ceph-mgr to load a module
-    from .module import Module, StandbyModule  # noqa: F401
 else:
     import logging
     logging.basicConfig(level=logging.DEBUG)
@@ -32,7 +30,12 @@ else:
     os.environ['PATH'] = '{}:{}'.format(os.path.abspath('../../../../build/bin'),
                                         os.environ['PATH'])
 
-    from tests import mock  # type: ignore
+    from tests import mock, mock_ceph_modules  # type: ignore
 
     mgr = mock.Mock()
     mgr.get_frontend_path.side_effect = lambda: os.path.abspath("./frontend/dist")
+
+    mock_ceph_modules()
+
+# DO NOT REMOVE: required for ceph-mgr to load a module
+from .module import Module, StandbyModule  # noqa: F401
